input validation with zod
