+++ packages/typed-api/runtime/error.ts
@@ -1,0 +1,6 @@
+export class TypedAPIError extends Error {
+    name = "TypedAPIError"
+    constructor(message: string, cause?: any) {
+        super(message, { cause })
+    }
+}
