+++ tests/fixtures/typed-api/src/pages/api/x.ts
@@ -1,7 +1,9 @@
 import { defineApiRoute } from "astro-typed-api/server"
+import z from "zod"
 
 export const POST = defineApiRoute({
-    fetch(input: { x: string }, context) {
-        return { y: input.x + "xyz" }
+    schema: z.object({ x: z.string(), y: z.number() }),
+    fetch({ x, y }, { locals, cookies }) {
+        return { z: x + "xyz" + y }
     }
 })
