--- tests/fixtures/typed-api/src/pages/x.ts
@@ -1,7 +1,0 @@
-import { defineApiRoute } from "astro-typed-api/server"
-
-export const POST = defineApiRoute({
-    fetch(input: { x: string }, context) {
-        return { y: input.x + "xyz" }
-    }
-})
